Rename package 'membership' to 'member' for better readability
