+++ src/test/java/edu/member/impl/UdpNetworkTest.java
@@ -1,11 +1,11 @@
-package edu.membership.impl;
+package edu.member.impl;
 
 import edu.common.api.Payload;
 import edu.common.impl.RemoteAddress;
 import edu.common.impl.UdpNetwork;
-import edu.membership.api.Member;
+import edu.member.api.Member;
 import edu.common.api.Network;
-import edu.membership.api.Message;
+import edu.member.api.Message;
 import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
